+++ p2panda-rs/src/operation/mod.rs
@@ -8,7 +8,13 @@ mod error;
 #[allow(clippy::module_inception)]
 mod operation;
 mod operation_encoded;
+mod operation_meta;
 
-pub use error::{OperationEncodedError, OperationError, OperationFieldsError};
-pub use operation::{Operation, OperationAction, OperationFields, OperationValue};
+pub use error::{
+    OperationEncodedError, OperationError, OperationFieldsError, OperationWithMetaError,
+};
+pub use operation::{
+    AsOperation, Operation, OperationAction, OperationFields, OperationValue, OperationVersion,
+};
 pub use operation_encoded::OperationEncoded;
+pub use operation_meta::OperationWithMeta;